requirement.txt
antlr4-python3-runtime==4.9.2
PySide2==5.15.2.1
shiboken2==5.15.2.1